Quantity mapped to Id within ProductTableRow
Controversial, but the quantity of the Product is held in the Redux
state tree rather than in local UI state. The basket has an items object
that contains hashmaps in the form of "id":"quantity" so as to flatten
the data-structure.
Styles are refactored for simplicity, removing classnames in favour of
elements